Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {FlatList,View,Text} from 'react-native'
+import {FlatList,View,Text,Alert} from 'react-native'
 import {ListItem} from 'react-native-elements'
 import { connect } from 'react-redux'
 import {baseUrl} from '../shared/baseUrl'
@@ -15,7 +15,24 @@ function Favorites(props){
             {
                 text:'Delete',
                 type:'delete',
-                onPress:()=>props.deleteFavorite(item.id)
+                onPress:()=>{
+                    Alert.alert(
+                        'Delet Favorite?',
+                        'Are you sure you wish to delete the favorite dish '+item.name+'?',
+                        [
+                            {
+                                text:'Cancel',
+                                onPress:()=>console.log(item.name+'Not Deleted'),
+                                style: 'cancel'
+                            },
+                            {
+                                text:'Okay',
+                                onPress:()=>props.deleteFavorite(item.id),
+                            }
+                        ],
+                        {cancelable:false}
+                    )
+                }
             }
         ]
         return(
